WIP refactor - flatten
