+++ src/controllers/survey-response.controller.ts
@@ -1,9 +1,9 @@
 import { v4 as uuid } from 'uuid';
 import { None, Option } from 'funfix';
-import { SurveyResponse } from './survey-response.entity';
-import { SurveyResponseRepository, SurveyId, SurveyResponseId } from './survey-response.repository';
-import { SurveyAnswer } from './survey-answer';
-import { SubmissionId } from '../submission/submission.types';
+import { SurveyResponse } from '../types/survey-response.entity';
+import { SurveyResponseRepository, SurveyId, SurveyResponseId } from '../types/survey.types';
+import { SurveyAnswer } from '../types/survey-answer';
+import { SubmissionId } from '../types/submission.types';
 import { Logger } from '@nestjs/common';
 
 export class SurveyResponseController {
