--- src/modules/auth/auth.module.ts
@@ -1,18 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { PassportModule } from '@nestjs/passport';
-import { JwtStrategy } from './jwt.strategy';
-import { JwtModule } from '@nestjs/jwt';
-import { jwtSecret } from './constants';
-// REMOVE - whole folder just json-token
-@Module({
-    imports: [
-        PassportModule,
-        JwtModule.register({
-            secret: jwtSecret,
-            signOptions: { expiresIn: '1 day' },
-        }),
-    ],
-    providers: [AuthService, JwtStrategy],
-})
-export class AuthModule {}
