--- src/modules/auth/constants.ts
@@ -1,1 +1,0 @@
-export const jwtSecret = process.env.JWT_SECRET || 'not-very-secret';
