--- src/packages/submission/index.ts
@@ -1,1 +1,0 @@
-// Nothing to see here!
