+++ src/schemas/survey.graphql
