+++ src/services/survey.service.ts
@@ -2,11 +2,11 @@ import { Injectable, OnModuleDestroy } from '@nestjs/common';
 import { Option, Some, None } from 'funfix';
 import * as Knex from 'knex';
 import { ConfigService } from '../modules/config/config.service';
-import { KnexSurveyResponseRepository } from '../modules/survey-adaptor/survey-response.repository';
-import { SurveyResponseController } from 'src/packages/survey/survey-response.controller';
-import { SurveyAnswer } from 'src/packages/survey/survey-answer';
-import { SurveyResponse } from 'src/packages/survey/survey-response.entity';
-import { SurveyId } from 'src/packages/survey/survey-response.repository';
+import { KnexSurveyResponseRepository } from '../repositories/survey-response.repository';
+import { SurveyResponseController } from 'src/controllers/survey-response.controller';
+import { SurveyAnswer } from 'src/types/survey-answer';
+import { SurveyResponse } from 'src/types/survey-response.entity';
+import { SurveyId } from 'src/repositories/survey-response.repository';
 import { SubmissionId } from '../types/submission.types';
 // REMOVE - MAYBE? PROBABLY!
 @Injectable()
