+++ src/types/survey-response.entity.test.ts
@@ -1,6 +1,6 @@
 import { v4 as uuid } from 'uuid';
 import { SurveyResponse } from './survey-response.entity';
-import { SurveyId, SurveyResponseId } from './survey-response.repository';
+import { SurveyId, SurveyResponseId } from '../repositories/survey-response.repository';
 import { SubmissionId } from '../submission/submission.types';
 import { Question } from './question';
 import { Answer } from './answer';
